requirements.txt
pillow
pytesseract
tkinter
